Move MockCheckPointStorage to the infrastructure folder.
