+++ sdk/eventhub/Azure.Messaging.EventHubs/samples/Infrastructure/MockCheckPointStorage.cs
@@ -5,8 +5,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Azure.Messaging.EventHubs.Processor;
 
-namespace Azure.Messaging.EventHubs.Processor
+namespace Azure.Messaging.EventHubs.Samples.Infrastructure
 {
     /// <summary>
     /// The EventProcessor relies on a <see cref="PartitionManager" /> to store checkpoints and handle partition
