+++ sdk/eventhub/Azure.Messaging.EventHubs/tests/EventProcessorClient/MockCheckPointStorageTests.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Azure.Messaging.EventHubs.Processor;
 using NUnit.Framework;
+using Azure.Messaging.EventHubs.Samples.Infrastructure;
 
 namespace Azure.Messaging.EventHubs.Tests
 {
